% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{from_json}
\alias{from_json}
\title{Parse json into lossless R representation}
\usage{
from_json(string)
}
\arguments{
\item{string}{A string to parse as json}
}
\value{
An object of class \code{tinyjson}, which is our lossless
representation of json data.
}
\description{
Parse a json string into a lossless R representation. Unlike
\code{jsonlite::fromJSON} this does not try and deserialise the JSON
into conventional R objects, but instead converts the data into a
representation that can be inspected, modified and rewritten with
\link{to_json} without any corruption caused by the ambiguity
of R's lack of a true scalar type.
}
\examples{
tinyjson::from_json('[1, 2, 3, 4]')
tinyjson::from_json('{"a": "value"}')
}
\seealso{
\link{from_json}, which does the inverse
transformation
}
