% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{to_json}
\alias{to_json}
\title{Convert json object to string}
\usage{
to_json(x)
}
\arguments{
\item{x}{A json object to convert}
}
\value{
A string
}
\description{
Format a json object read in via \link{from_json} into a
string. Importantly, and unlike \code{jsonlite::toJSON} this does not
work for arbitrary R objects.
}
\examples{
# Lossless json -> R -> json roundtrip of arrays:
tinyjson::to_json(tinyjson::from_json('1'))
tinyjson::to_json(tinyjson::from_json('[1]'))
tinyjson::to_json(tinyjson::from_json('[1, 2, 3, 4]'))
}
\seealso{
\link{from_json}, which does the inverse
transformation
}
